move more chat logging code into chatlog.py
